Verified 1) model_dir output siloing and 2) Py3.7
1) The model_dir parameter can indeed be used for siloing output
from multiple instances of the SuperLearner.

2) Python 3.7 does NOT solve the string decode error.  I have no idea
why this has not been an issue in the past.  Although the fix is
easy (simply remove ".decode('latin1')" from a Python file) this
intervention does not lend itself well to automated deployment.
Consider more recent versions of Scikit-Learn to sidestep this
issue as well as allow for non-negative least squares in
LinearRegression(positive=True) (only since sklearn 0.24).