Removed set -e to prevent crashes on failure
Some git commands will fail, e.g. if an already running cluster
attempts to git clone a repo that already exists.  There is probably
a more sophisticated way around this in git, but I don't know it
yet.  For now, simply remove set -e so that if a git clone
operation fails, execution simply moves on to the next command
instead of crashing the whole workflow. This also applies to the
situation if a branch does not already exist.  As they are now,
the git commands here will clone repos/create branches if not
already present or fail (but not crash whole workflow) if the
state of those repos/branches does not match expectations of the
command. This approach allows for reusing a "warm" cluster
(i.e. keeping the head node up and running the workflow several
times without needing to clone everything all over again).