Compress req.txt in model archive
This reduces the size of the archive (a tiny bit) but also
sidesteps the GH Dependabot check for issues with
requirements.txt. The HPC code that is using the Conda environment
is ephemeral and siloed; it is not "production" code (as in a
server's code interacting with the outside world) so the DependaBot
checks/warnings do not apply to this case.