Force monitor to wait first, then check
Squeue may not register a running job instantly, but ssh sbatch
finishes very quickly.  So, reorder the monitor so that it waits
first and then checks the number of lines in squeue (instead of
first checking number of lines and then waiting).  This allows
some time for squeue and the SLURM infrastructure to register
the new job.