Exit SSH connections for key forwarding
If a background/interactive ssh session is open and the user tries
to ssh to the same node with key forwarding (-A, e.g. for an ssh-
authenticated git operation), the command will fail. Test the
very simple approach of announcing to all users that ssh -O exit
will be called (which may interrupt a session), close the
background/interactive session, and then use ssh -A for the necessary
git command. This is clunky and need to find the root cause of
the issue and implement a less intrusive solution.