Working function/inverse pair
Key issue was to ensure that outputs of MinMaxScaler
in function and inputs to MinMaxScaler in inverse
are between 0 and 1. This applies really only to the
INVERSE - likely because the ML models themselves are
not conscious of the need to produce MinMaxScaled
output (between 0 and 1), especially when they are
given a very wide range of hyperparameters that can
cause very strange model behavior in certain
combinations.